work on CLC

+++ eclipse-workspace/cpu6502/src/main/java/cpu6502/main/components/ArithmeticLogicUnit.java
@@ -2,8 +2,8 @@ package cpu6502.main.components;
 
 public class ArithmeticLogicUnit {
 	
-	public byte aInputRegister;
-	public byte bInputRegister;
+	public int aInputRegister;
+	public int bInputRegister;
 	
 	public boolean sums;
 	public boolean ands;
